extract ButtonColor class into separate file.
