+++ lib/src/utils/auth_button_color.dart
@@ -1,0 +1,18 @@
+// Copyright (c) 2021 Talat El Beick. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+
+@immutable
+class ButtonColor {
+  final Color color;
+  final Color? onSecondaryIcon;
+  final Color? onSecondaryButton;
+
+  const ButtonColor(
+    this.color, {
+    this.onSecondaryIcon,
+    this.onSecondaryButton,
+  });
+}
